Add files via upload
Spent a lot of time on PowerBi.  Started off as a separate analysis but found the visualizations a great way to show my findings or explain them.  I still need to polish up the PowerBi files, so I could use for a standalone presentation, because the interaction feature is great!