My Different Notebooks for my Steel Project
Some of these notebooks are ugly and  was using because my memory was getting sucked up and I didn't put in the correct code to reduce the memory eating analysis.    Sometimes, I get too excited about the next analysis and model and forget to take care of maintenance.  It was a fun journey for me, even with the multiple notebooks.